feat:cambio en el modelo de assessment #157

+++ assessment/models.py
@@ -11,7 +11,7 @@ class Assessment(models.Model):
     
     id = models.PositiveIntegerField(primary_key=True, default=random_id, editable=False)
 
-    stars = models.IntegerField(validators=[MinValueValidator(1), MaxValueValidator(5)])
+    stars = models.FloatField(validators=[MinValueValidator(1), MaxValueValidator(5)])
     equipment = models.ForeignKey(Equipment, on_delete=models.CASCADE)
     client = models.ForeignKey(Client, on_delete=models.CASCADE)
